comparing with c++ is nice
added comparison details

Update README.md

Update README.md

typo

Update COMP_DETAILS.md

Update README.md

touch up compare_cpp

touch up compare_cpp again

touch up again

clean out old files

no image

Update README.md

fixed link

Update README.md

actually fixed link

Update README.md

Update README.md

Update README.md

Update README.md

rename

added tutorial for vnlb

Update COMPARE.md

Update README.md

typos

format

rephrase

rephrase

only 5 example davis frames

Update VNLB.md

Update VNLB.md

Update VNLB.md

Update VNLB.md

Update COMPARE.md

Update README.md

Update COMPARE.md

Update README.md

Update COMPARE.md

mod data; flow2burst; cmake note

Update VNLB.md

Update VNLB.md

Update COMPARE.md

Update COMPARE.md


Former-commit-id: 138756792dceeccfc73e0f8feb87128baf604c68

+++ pylib/utils.py
@@ -14,8 +14,7 @@ def ndarray_ctg_dtype(ndarray,dtype,verbose):
     return ndarray
 
 def rgb2bw(burst):
-    # burst_bw = .299 * burst[:,2] + .587 * burst[:,1] + .114 * burst[:,0]
-    # burst_bw = burst_bw[:,None].copy()
+    burst = burst.astype(np.float32)
     burst_bw = []
     for t in range(burst.shape[0]):
         frame = burst[t]
@@ -25,9 +24,6 @@ def rgb2bw(burst):
         frame = rearrange(frame,'h w -> 1 h w')
         burst_bw.append(frame)
     burst_bw = np.stack(burst_bw).copy()
-    # import torch
-    # import torchvision.utils as tvUtils
-    # tvUtils.save_image(torch.FloatTensor(burst_bw)/255.,"burst_bw_inner.png")
     return burst_bw
 
 def compute_psnrs(img1,img2,imax=255.):
